# cmd/api/docs/swagger.yaml
definitions:
...
  models.OTPData:
    properties:
      phone:
        type: string
    required:
      - phone
    type: object
...
  models.Order:
    properties:
      address_id:
        type: integer
      coupon_id:
        type: integer
      payment_id:
        type: integer
      user_id:
        type: integer
    type: object
...
  models.VerifyData:
    properties:
      code:
        type: string
      phone:
        type: string
    required:
      - code
      - phone
    type: object
  response.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      status_code:
        type: integer
    type: object
info:
  contact: {}
paths:
...
  /admin/orders:
    get:
      consumes:
        - application/json
      description: Admin can view the orders according to status
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/response.Response"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/response.Response"
      security:
        - Bearer: []
      summary: Admin Orders
      tags:
        - Admin
  /admin/orders/edit/status:
    put:
      consumes:
        - application/json
      description: Admin can change the status of the order
      parameters:
        - description: id
          in: query
          name: id
          required: true
          type: string
        - description: status
          in: query
          name: status
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/response.Response"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/response.Response"
      security:
        - Bearer: []
      summary: Update Order Status
      tags:
        - Admin
...
  /users/check-out/order:
    post:
      consumes:
        - application/json
      description: user can order the items that currently in cart
      parameters:
        - description: id
          in: body
          name: order
          required: true
          schema:
            $ref: "#/definitions/models.Order"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/response.Response"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/response.Response"
      security:
        - Bearer: []
      summary: Order Now
      tags:
        - User
...
  /users/otplogin:
    post:
      consumes:
        - application/json
      description: OTP login send otp
      parameters:
        - description: otp-data
          in: body
          name: otp
          required: true
          schema:
            $ref: "#/definitions/models.OTPData"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/response.Response"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/response.Response"
      summary: Send OTP
      tags:
        - User
...
  /users/profile/orders:
    delete:
      consumes:
        - application/json
      description: user can cancel the orders
      parameters:
        - description: id
          in: query
          name: id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/response.Response"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/response.Response"
      security:
        - Bearer: []
      summary: Order Cancel
      tags:
        - User
    get:
      consumes:
        - application/json
      description: user can view the details of the orders
      parameters:
        - description: id
          in: query
          name: id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/response.Response"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/response.Response"
      security:
        - Bearer: []
      summary: Get Orders
      tags:
        - User
  /users/profile/orders/return:
    put:
      consumes:
        - application/json
      description:
        user can return the ordered products which is already delivered
        and then get the amount fot that particular purchase back in their wallet
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/response.Response"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/response.Response"
      security:
        - Bearer: []
      summary: Return Order
      tags:
        - User
...
  /users/verifyotp:
    post:
      consumes:
        - application/json
      description: OTP login verify otp
      parameters:
        - description: otp-verify
          in: body
          name: otp
          required: true
          schema:
            $ref: "#/definitions/models.VerifyData"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/response.Response"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/response.Response"
      summary: Verify OTP
      tags:
        - User
swagger: "2.0"